deploy/data_migration/requirements.txt
pandas
psycopg2-binary
pymongo